거리 및 HP 기준 정렬 추가
최저 HP 타겟팅 기능 작업 중

+++ Assets/Scripts/Authoring/FindTargetAuthoring.cs
@@ -56,7 +56,7 @@ public struct FindTarget : IComponentData
 {
     public float MinDistance;
     public float MaxDistance;
-    public float MaxTargets;
+    public int MaxTargets;
     public EUnitType TargetingUnitType;
     public float Timer;
     public float TimerMax;
